Added worker metrics. Added metrics config support. Refactor code to split in files. Bump pond version to 1.8.3 (patch). Clean up the code.
