+++ app/cmd/cli/root.go
@@ -2,6 +2,7 @@ package cli
 
 import (
 	"fmt"
+	"github.com/pokt-network/pocket-core/app/cmd/rpc/mesh"
 	"log"
 	"os"
 	"os/signal"
@@ -139,7 +140,7 @@ var meshCmd = &cobra.Command{
 
 func startMesh(cmd *cobra.Command, args []string) {
 	app.InitMeshConfig(datadir)
-	rpc.StartMeshRPC(simulateRelay)
+	mesh.StartMeshRPC(simulateRelay)
 }
 
 // resetCmd represents the reset command
