+++ app/cmd/rpc/rpc_test.go
@@ -1401,7 +1401,7 @@ func TestRPC_SimRelay(t *testing.T) {
 		Data:   expectedRequest,
 		Method: "POST",
 	}
-	simParams := simRelayParams{
+	simParams := SimRelayParams{
 		RelayNetworkID: dummyChainsHash,
 		Payload:        payload,
 	}
